ALSA: timer: Fix race among timer ioctls
Issue: SCP6-766

1/1 [
Author: Takashi Iwai
Email: tiwai@suse.de
Subject: ALSA: timer: Fix race among timer ioctls
Date: Wed, 13 Jan 2016 17:48:01 +0100

commit af368027a49a751d6ff4ee9e3f9961f35bb4fede upstream

ALSA timer ioctls have an open race and this may lead to a
use-after-free of timer instance object.  A simplistic fix is to make
each ioctl exclusive.  We have already tread_sem for controlling the
tread, and extend this as a global mutex to be applied to each ioctl.

The downside is, of course, the worse concurrency.  But these ioctls
aren't to be parallel accessible, in anyway, so it should be fine to
serialize there.

Reported-by: Dmitry Vyukov <dvyukov@google.com>
Tested-by: Dmitry Vyukov <dvyukov@google.com>
Cc: <stable@vger.kernel.org>
Signed-off-by: Takashi Iwai <tiwai@suse.de>
Signed-off-by: Yue Tao <Yue.Tao@windriver.com>
]

Signed-off-by: Yue Tao <yue.tao@windriver.com>