rtmutex: Handle non enqueued waiters gracefully
Issue: LIN7-6765

1/1 [
Author: "Junxian.Xiao"
Email: Junxian.Xiao@windriver.com
Subject: rtmutex: Handle non enqueued waiters gracefully
Date: Mon, 5 Sep 2016 04:36:51 +0800

Commit 2ae5b87eab12aa5773a4da028c2d386bfd4d9dc5 upstream

Yimin debugged that in case of a PI wakeup in progress when
rt_mutex_start_proxy_lock() calls task_blocks_on_rt_mutex() the latter
returns -EAGAIN and in consequence the remove_waiter() call runs into
a BUG_ON() because there is nothing to remove.

Guard it with rt_mutex_has_waiters(). This is a quick fix which is
easy to backport. The proper fix is to have a central check in
remove_waiter() so we can call it unconditionally.

Reported-and-debugged-by: Yimin Deng <yimin11.deng@gmail.com>
Signed-off-by: Thomas Gleixner <tglx@linutronix.de>
Cc: stable-rt@vger.kernel.org
Signed-off-by: Junxian.Xiao <Junxian.Xiao@windriver.com>
]

Signed-off-by: Yue Tao <Yue.Tao@windriver.com>