ovl: fix dentry leak for default_permissions
Issue: LIN8-3569, LIN8-3526

1/3 [
Author: Miklos Szeredi
Email: miklos@szeredi.hu
Subject: ovl: default permissions
Date: Mon, 12 Oct 2015 17:11:44 +0200

commit 8d3095f4ad47ac409440a0ba1c80e13519ff867d upstream

Add mount option "default_permissions" to alter the way permissions are
calculated.

Without this option and prior to this patch permissions were calculated by
underlying lower or upper filesystem.

With this option the permissions are calculated by overlayfs based on the
file owner, group and mode bits.

This has significance for example when a read-only exported NFS filesystem
is used as a lower layer.  In this case the underlying NFS filesystem will
reply with EROFS, in which case all we know is that the filesystem is
read-only.  But that's not what we are interested in, we are interested in
whether the access would be allowed if the filesystem wasn't read-only; the
server doesn't tell us that, and would need updating at various levels,
which doesn't seem practicable.

Signed-off-by: Miklos Szeredi <miklos@szeredi.hu>
Signed-off-by: Yue Tao <Yue.Tao@windriver.com>
]

2/3 [
Author: Stephen Hemminger
Email: stephen@networkplumber.org
Subject: ovl: move super block magic number to magic.h
Date: Wed, 4 Nov 2015 10:59:52 -0800

commit 257f871993474e2bde6c497b54022c362cf398e1 upstream

The overlayfs file system is not recognized by programs
like tail because the magic number is not in standard header location.

Move it so that the value will propagate on for the GNU library
and utilities. Needs to go in the fstatfs manual page as well.

Signed-off-by: Stephen Hemminger <stephen@networkplumber.org>
Signed-off-by: Miklos Szeredi <miklos@szeredi.hu>
Signed-off-by: Yue Tao <Yue.Tao@windriver.com>
]

3/3 [
Author: Miklos Szeredi
Email: mszeredi@redhat.com
Subject: ovl: fix dentry leak for default_permissions
Date: Wed, 29 Jun 2016 08:26:59 +0200

commit a4859d75944a726533ab86d24bb5ffd1b2b7d6cc upstream

When using the 'default_permissions' mount option, ovl_permission() on
non-directories was missing a dput(alias), resulting in "BUG Dentry still
in use".

Signed-off-by: Miklos Szeredi <mszeredi@redhat.com>
Fixes: 8d3095f4ad47 ("ovl: default permissions")
Cc: <stable@vger.kernel.org> # v4.5+
Signed-off-by: Yue Tao <Yue.Tao@windriver.com>
]

Signed-off-by: Yue Tao <Yue.Tao@windriver.com>