rtmutex: Fix up failed merge update
Issue: LIN6-11742

1/1 [
Author: "Steven Rostedt (Red Hat)"
Email: rostedt@goodmis.org
Subject: rtmutex: Fix up failed merge update
Date: Thu, 30 Oct 2014 15:52:34 -0400

commit d998a5df1d48dc4df6b44a2b5051aacecf76186a upstream

The -rt patch replaced the wake_up_prcocess() in rtmutex.c with
rt_mutex_wake_waiter(). When merging 3.10.49, I did not remove
the wake_up_process() and left it there after the rt_mutex_wake_waiter().

This passed the tests, but is not correct. There's a tight race where
a task sets its state to TASK_UNINTERRUPTIBLE and grabs a spin_lock
that was converted into an rtmutex in -rt. The rtmutex must keep the
tasks state otherwise it fail to schedule after releasing the spin lock
and calling schedule.

In most cases, this will not cause any harm, but there's a 1 in a million
chance that it will. The tests survived because it's hard to hit this
bug.

As I didn't catch this after I pushed the changes to kernel.org, I'm
making this fix now as a separate -rt release.

Signed-off-by: Steven Rostedt <rostedt@goodmis.org>
Signed-off-by: Zhenbo Gao <zhenbo.gao@windriver.com>
]

Signed-off-by: Yue Tao <yue.tao@windriver.com>