uart: set the interrupt trigger type to edge in dts
Issue: LIN8-4847

1/1 [
Author: "Junxian.Xiao"
Email: Junxian.Xiao@windriver.com
Subject: uart: set the interrupt trigger type to edge in dts
Date: Fri, 21 Oct 2016 13:42:38 +0800

In the original dst settings, the uart0/1 are sharing the
same interrupt pin and both trigger by is high level.
When ttyS0 as console, and also try to use ttyS1, there is
the folowing call trace:

    irq 18: nobody cared (try booting with the "irqpoll" option)
    ...
    handlers:
    [<ffff8000000ed258>] irq_default_primary_handler threaded
    [<ffff8000004d8730>] serial8250_interrupt
    Disabling IRQ #18

This only happens in preempt-rt kernel, which using the thread
as the irq bottom part, more possible that the handler cannot
the sharing interrupt at the same ttime. Just set the trigger
type to edge to avoid timing conflict.

Signed-off-by: Junxian.Xiao <Junxian.Xiao@windriver.com>
]

Signed-off-by: Yue Tao <Yue.Tao@windriver.com>