disable ZYNQMP_PM_DOMAINS and drop muiltmedia and PCIe out by default.
Generally speaking, there is a related-TRD firmware images for every
new release according to Xilinx website, but TRD for 2016.3 is still
not shown up, so I split these two featues out by default since there
are no-related TRD to be validated.

The ZYNQMP_PM_DOMAINS related hardware PL is still not vailable, which
means the related codes are not validated, even if in SDK, so we disable
it by default.

Signed-off-by: Zumeng Chen <zumeng.chen@windriver.com>