netfilter: add option required by lxc-checkconfig
LIN9-2138

One of the main users of the netfilter feature is lxc which has
requirements for some specific kernel configuration options, available
for audit using lxc-checkconfig. In order to properly support
container networking via lxcbr0 we have to ensure we enable the
required kconfigs to enable full NAT control, for masquerade and
forwarding... The added configs are required to pass the
lxc-checkconfig audit but are also useful in most cases where
netfilter via iptables or nft is in use.

Signed-off-by: Mark Asselstine <mark.asselstine@windriver.com>