packet: fix race condition in packet_set_ring
Issue: SCP7-481

1/1 [
Author: Philip Pettersson
Email: philip.pettersson@gmail.com
Subject: packet: fix race condition in packet_set_ring
Date: Wed, 30 Nov 2016 10:07:55 +0900

commit 84ac7260236a49c79eede91617700174c2c19b0cq upstream

When packet_set_ring creates a ring buffer it will initialize a
struct timer_list if the packet version is TPACKET_V3. This value
can then be raced by a different thread calling setsockopt to
set the version to TPACKET_V1 before packet_set_ring has finished.

This leads to a use-after-free on a function pointer in the
struct timer_list when the socket is closed as the previously
initialized timer will not be deleted.

The bug is fixed by taking lock_sock(sk) in packet_setsockopt when
changing the packet version while also taking the lock at the start
of packet_set_ring.

Fixes: f6fb8f100b80 ("af-packet: TPACKET_V3 flexible buffer implementation.")
Signed-off-by: Philip Pettersson <philip.pettersson@gmail.com>
Signed-off-by: Eric Dumazet <edumazet@google.com>
Signed-off-by: David S. Miller <davem@davemloft.net>
Signed-off-by: Zhang Xiao <xiao.zhang@windriver.com>
]

Signed-off-by: Yue Tao <Yue.Tao@windriver.com>