standard/base: export from and resync to fast forward tree.
Nothing has been added to the WR9 kernel cache since the final
treegen took place.  But we want to have an rt rebase branch as
we've done in yocto, which also means we need a standard/rebase
branch implicitly.

As of this import, standard/base from the ff tree, and a treegen
from the kernel-cache are source equivalent as of this commit:

  commit 79b24818b3aa573658ba3d4f8fb13b565161fe4a
  Author: Jianchuan Wang <jianchuan.wang@windriver.com>
  Date:   Fri Dec 9 10:58:01 2016 +0800

    lttng: Fix recursive dependency between LTTng and TRACEPOINTS

Further updates for -rt specific patches are expected next.

Signed-off-by: Paul Gortmaker <paul.gortmaker@windriver.com>