rt: drop old 3.x patch
Issue: LIN9-2658

The fix was designed for early 3.x-rt releases, and even then for
a fixed subset, since the issue was resolved in later stable updates
in a different way.

While having it present should not present any runtime issues, there
is no point in deviating from the main -rt release with this.

Signed-off-by: Paul Gortmaker <paul.gortmaker@windriver.com>