posix_acl: Clear SGID bit when setting file permissions
Issue: LIN5-22058

1/2 [
Author: "Eric W. Biederman"
Email: ebiederm@xmission.com
Subject: userns: Replace the hard to write inode_userns with inode_capable.
Date: Mon, 14 Nov 2011 16:24:06 -0800

commit 1a48e2ac034d47ed843081c4523b63c46b46888b upstream

This represents a change in strategy of how to handle user namespaces.
Instead of tagging everything explicitly with a user namespace and bulking
up all of the comparisons of uids and gids in the kernel,  all uids and gids
in use will have a mapping to a flat kuid and kgid spaces respectively.  This
allows much more of the existing logic to be preserved and in general
allows for faster code.

In this new and improved world we allow someone to utiliize capabilities
over an inode if the inodes owner mapps into the capabilities holders user
namespace and the user has capabilities in their user namespace.  Which
is simple and efficient.

Moving the fs uid comparisons to be comparisons in a flat kuid space
follows in later patches, something that is only significant if you
are using user namespaces.

Acked-by: Serge Hallyn <serge.hallyn@canonical.com>
Signed-off-by: Eric W. Biederman <ebiederm@xmission.com>
Signed-off-by: Alexandru Moise <alexandru.moise@windriver.com>
]

2/2 [
Author: Alexandru Moise
Email: alexandru.moise@windriver.com
Subject: posix_acl: Clear SGID bit when setting file permissions
Date: Mon, 12 Dec 2016 15:09:24 +0200

commit 073931017b49d9458aa351605b43a7e34598caef upstream

When file permissions are modified via chmod(2) and the user is not in
the owning group or capable of CAP_FSETID, the setgid bit is cleared in
inode_change_ok().  Setting a POSIX ACL via setxattr(2) sets the file
permissions as well as the new ACL, but doesn't clear the setgid bit in
a similar way; this allows to bypass the check in chmod(2).  Fix that.

References: CVE-2016-7097
Reviewed-by: Christoph Hellwig <hch@lst.de>
Reviewed-by: Jeff Layton <jlayton@redhat.com>
Signed-off-by: Jan Kara <jack@suse.cz>
Signed-off-by: Andreas Gruenbacher <agruenba@redhat.com>
Signed-off-by: Alexandru Moise <alexandru.moise@windriver.com>
]

Signed-off-by: Yue Tao <yue.tao@windriver.com>