tty: Serialize function receive_room
Issue: LIN7-7247

1/1 [
Author: Qi Hou
Email: qi.hou@windriver.com
Subject: tty: Serialize function receive_room
Date: Wed, 28 Dec 2016 08:34:17 +0800

There is a read buffer in pty device and it has two variables to
check and judge the status of this read buffer. "read_cnt" shows
how many bytes can be read from this read buffer; "receive_room"
shows how many bytes are still free and can be write into the read
buffer. "read_cnt" is 0 means the buffer is empty while the 0
"receve_room" means the buffer is  full. The "receive_room" was
set based on "read_cnt" with function "receive_room" by both the
"consumer" and "producer" of this read buffer. While the very
function was not serialized that may cause race condition in
preempt kernel. That race condition may cause both the raed_cnt and
receive_room is 0 thus block both consumer and producer for ever.

Without serialization, preemption on this function may
cause the read_cnt and receive_room are both set to
zero. This will block both producer and consumer of the
read_buf. Serialize it with a new created spinlock.

Signed-off-by: Qi Hou <qi.hou@windriver.com>
Signed-off-by: Zhang Xiao <xiao.zhang@windriver.com>
]

Signed-off-by: Yue Tao <Yue.Tao@windriver.com>