posix_acl: Clear SGID bit when setting file permissions
Issue: SCP7-460

1/1 [
Author: Alexandru Moise
Email: alexandru.moise@windriver.com
Subject: posix_acl: Clear SGID bit when setting file permissions
Date: Tue, 13 Dec 2016 20:43:20 +0200

commit 073931017b49d9458aa351605b43a7e34598caef upstream

When file permissions are modified via chmod(2) and the user is not in
the owning group or capable of CAP_FSETID, the setgid bit is cleared in
inode_change_ok().  Setting a POSIX ACL via setxattr(2) sets the file
permissions as well as the new ACL, but doesn't clear the setgid bit in
a similar way; this allows to bypass the check in chmod(2).  Fix that.

References: CVE-2016-7097
Reviewed-by: Christoph Hellwig <hch@lst.de>
Reviewed-by: Jeff Layton <jlayton@redhat.com>
Signed-off-by: Jan Kara <jack@suse.cz>
Signed-off-by: Andreas Gruenbacher <agruenba@redhat.com>
Signed-off-by: Alexandru Moise <alexandru.moise@windriver.com>
]

Signed-off-by: Yue Tao <Yue.Tao@windriver.com>