netlink: Fix dump skb leak/double free
Issue: LIN6-12190

1/1 [
Author: Herbert Xu
Email: herbert@gondor.apana.org.au
Subject: netlink: Fix dump skb leak/double free
Date: Mon, 9 Jan 2017 15:38:42 +0800

commit 92964c79b357efd980812c4de5c1fd2ec8bb5520 upstream

When we free cb->skb after a dump, we do it after releasing the
lock.  This means that a new dump could have started in the time
being and we'll end up freeing their skb instead of ours.

This patch saves the skb and module before we unlock so we free
the right memory.

Fixes: 16b304f3404f ("netlink: Eliminate kmalloc in netlink dump operation.")
Reported-by: Baozeng Ding <sploving1@gmail.com>
Signed-off-by: Herbert Xu <herbert@gondor.apana.org.au>
Acked-by: Cong Wang <xiyou.wangcong@gmail.com>
Signed-off-by: David S. Miller <davem@davemloft.net>
[Make some changes to apply it to our tree]
Signed-off-by: Wenlin Kang <wenlin.kang@windriver.com>
]

Signed-off-by: Yue Tao <yue.tao@windriver.com>