crypto: mcryptd - Check mcryptd algorithm compatibility
Issue: LIN8-5699

1/1 [
Author: tim
Email: tim.c.chen@linux.intel.com
Subject: crypto: mcryptd - Check mcryptd algorithm compatibility
Date: Mon, 5 Dec 2016 11:46:31 -0800

commit 48a992727d82cb7db076fa15d372178743b1f4cd upstream

Algorithms not compatible with mcryptd could be spawned by mcryptd
with a direct crypto_alloc_tfm invocation using a "mcryptd(alg)" name
construct.  This causes mcryptd to crash the kernel if an arbitrary
"alg" is incompatible and not intended to be used with mcryptd.  It is
an issue if AF_ALG tries to spawn mcryptd(alg) to expose it externally.
But such algorithms must be used internally and not be exposed.

We added a check to enforce that only internal algorithms are allowed
with mcryptd at the time mcryptd is spawning an algorithm.

Link: http://marc.info/?l=linux-crypto-vger&m=148063683310477&w=2
Cc: stable@vger.kernel.org
Reported-by: Mikulas Patocka <mpatocka@redhat.com>
Signed-off-by: Tim Chen <tim.c.chen@linux.intel.com>
Signed-off-by: Herbert Xu <herbert@gondor.apana.org.au>
Signed-off-by: Qi Hou <qi.hou@windriver.com>
]

Signed-off-by: Yue Tao <Yue.Tao@windriver.com>