KVM: x86: fix emulation of "MOV SS, null selector"
Issue: LIN8-5714

1/1 [
Author: Paolo Bonzini
Email: pbonzini@redhat.com
Subject: KVM: x86: fix emulation of "MOV SS, null selector"
Date: Thu, 12 Jan 2017 15:02:32 +0100

commit 33ab91103b3415e12457e3104f0e4517ce12d0f3 upstream

This is CVE-2017-2583.  On Intel this causes a failed vmentry because
SS's type is neither 3 nor 7 (even though the manual says this check is
only done for usable SS, and the dmesg splat says that SS is unusable!).
On AMD it's worse: svm.c is confused and sets CPL to 0 in the vmcb.

The fix fabricates a data segment descriptor when SS is set to a null
selector, so that CPL and SS.DPL are set correctly in the VMCS/vmcb.
Furthermore, only allow setting SS to a NULL selector if SS.RPL < 3;
this in turn ensures CPL < 3 because RPL must be equal to CPL.

Thanks to Andy Lutomirski and Willy Tarreau for help in analyzing
the bug and deciphering the manuals.

Reported-by: Xiaohan Zhang <zhangxiaohan1@huawei.com>
Fixes: 79d5b4c3cd809c770d4bf9812635647016c56011
Cc: stable@nongnu.org
Signed-off-by: Paolo Bonzini <pbonzini@redhat.com>
Signed-off-by: Haiqing Bai <Haiqing.Bai@windriver.com>
]

Signed-off-by: Yue Tao <Yue.Tao@windriver.com>