ipv4: keep skb->dst around in presence of IP options
Issue: LIN5-22575

1/1 [
Author: Eric Dumazet
Email: edumazet@google.com
Subject: ipv4: keep skb->dst around in presence of IP options
Date: Sat, 4 Feb 2017 11:16:52 -0800

commit 34b2cef20f19c87999fff3da4071e66937db9644 upstream

Andrey Konovalov got crashes in __ip_options_echo() when a NULL skb->dst
is accessed.

ipv4_pktinfo_prepare() should not drop the dst if (evil) IP options
are present.

We could refine the test to the presence of ts_needtime or srr,
but IP options are not often used, so let's be conservative.

Thanks to syzkaller team for finding this bug.

Fixes: d826eb14ecef ("ipv4: PKTINFO doesnt need dst reference")
Signed-off-by: Eric Dumazet <edumazet@google.com>
Reported-by: Andrey Konovalov <andreyknvl@google.com>
Signed-off-by: David S. Miller <davem@davemloft.net>
Signed-off-by: Qi Hou <qi.hou@windriver.com>
]

Signed-off-by: Yue Tao <yue.tao@windriver.com>