powerpc/mm: any thread in one core can be the first to setup TLB1
Issue: LIN8-5179

1/1 [
Author: chenhui zhao
Email: chenhui.zhao@freescale.com
Subject: powerpc/mm: any thread in one core can be the first to setup TLB1
Date: Thu, 24 Dec 2015 08:39:57 +0800

On e6500, in the case of cpu hotplug, either thread in one core
may be the first thread initilzing the TLB1. The subsequent threads
must not setup it again.

The code is derived from the comment of Scott Wood.

Signed-off-by: Chenhui Zhao <chenhui.zhao@freescale.com>
Signed-off-by: Scott Wood <oss@buserror.net>
Signed-off-by: Catalin Enache <catalin.enache@windriver.com>
Signed-off-by: Qi Hou <qi.hou@windriver.com>
]

Signed-off-by: Yue Tao <Yue.Tao@windriver.com>