drm/vmwgfx: fix integer overflow in vmw_surface_define_ioctl()
Issue: LIN9-3929

1/1 [
Author: Li Qiang
Email: liq3ea@gmail.com
Subject: drm/vmwgfx: fix integer overflow in vmw_surface_define_ioctl()
Date: Mon, 27 Mar 2017 20:10:53 -0700

commit e7e11f99564222d82f0ce84bd521e57d78a6b678 upstream

In vmw_surface_define_ioctl(), the 'num_sizes' is the sum of the
'req->mip_levels' array. This array can be assigned any value from
the user space. As both the 'num_sizes' and the array is uint32_t,
it is easy to make 'num_sizes' overflow. The later 'mip_levels' is
used as the loop count. This can lead an oob write. Add the check of
'req->mip_levels' to avoid this.

Cc: <stable@vger.kernel.org>
Signed-off-by: Li Qiang <liqiang6-s@360.cn>
Reviewed-by: Thomas Hellstrom <thellstrom@vmware.com>
Signed-off-by: Qi Hou <qi.hou@windriver.com>
]

Signed-off-by: Yue Tao <Yue.Tao@windriver.com>