fs: limit filesystem stacking depth
Issue: LIN5-22870

1/1 [
Author: Miklos Szeredi
Email: mszeredi@suse.cz
Subject: fs: limit filesystem stacking depth
Date: Tue, 18 Apr 2017 15:09:29 +0800

commit 69c433ed2ecd2d3264efd7afec4439524b319121 upstream

Add a simple read-only counter to super_block that indicates how deep
this
is in the stack of filesystems.  Previously ecryptfs was the only
stackable
filesystem and it explicitly disallowed multiple layers of itself.

Overlayfs, however, can be stacked recursively and also may be stacked
on top of ecryptfs or vice versa.

To limit the kernel stack usage we must limit the depth of the
filesystem stack.  Initially the limit is set to 2.

Signed-off-by: Miklos Szeredi <mszeredi@suse.cz>
Signed-off-by: Hu <yadi.hu@windriver.com>
]

Signed-off-by: Yue Tao <yue.tao@windriver.com>